Allow player selection via buttons vs text prompt
This commit modifies the orginal functions to take input from
buttons as opposed to a textbox. This vastly improves the user
experience as the previous method was tedious and only used due
to my limited knowledge at the time of of writing the initial code.
Changes have also been made to the CSS to improve aesthetics.